requirements.txt
beautifulsoup4==4.12.2
Django==4.2.4
django-admin-sortable2==2.1.9
django-bootstrap5==23.3
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-modal-forms==3.0.4
django-ckeditor==6.7.0
django-colorful==1.3
icalendar==5.0.7
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4